refactor: FetchPost 로직 분리
